DPR2-666 separate RedshiftDataApiRepository from ConfiguredApiRepository
