+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/RedshiftDataApiRepository.kt
@@ -1,0 +1,53 @@
+package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data
+
+import org.slf4j.LoggerFactory
+import org.springframework.stereotype.Service
+import software.amazon.awssdk.services.redshiftdata.RedshiftDataClient
+import software.amazon.awssdk.services.redshiftdata.model.ExecuteStatementRequest
+import software.amazon.awssdk.services.redshiftdata.model.ExecuteStatementResponse
+import software.amazon.awssdk.services.redshiftdata.model.SqlParameter
+
+@Service
+class RedshiftDataApiRepository(
+  val redshiftDataClient: RedshiftDataClient,
+  val executeStatementRequestBuilder: ExecuteStatementRequest.Builder,
+) : RepositoryHelper() {
+
+  companion object {
+    private val log = LoggerFactory.getLogger(this::class.java)
+  }
+  fun executeQueryAsync(
+    query: String,
+    filters: List<ConfiguredApiRepository.Filter>,
+    sortColumn: String?,
+    sortedAsc: Boolean,
+    reportId: String,
+    policyEngineResult: String,
+    dynamicFilterFieldId: String? = null,
+    dataSourceName: String,
+  ): String {
+    val statementRequest: ExecuteStatementRequest = executeStatementRequestBuilder
+      .sql(
+        buildFinalQuery(
+          buildReportQuery(query),
+          buildPolicyQuery(policyEngineResult),
+          buildFiltersQuery(filters),
+          buildFinalStageQuery(dynamicFilterFieldId, sortColumn, sortedAsc),
+        ),
+      )
+      .parameters(buildQueryParams(filters))
+      .build()
+
+    val response: ExecuteStatementResponse = redshiftDataClient.executeStatement(statementRequest)
+    log.info("Execution ID: {}", response.id())
+    return response.id()
+  }
+
+  fun buildQueryParams(filters: List<ConfiguredApiRepository.Filter>): List<SqlParameter> {
+    val sqlParams: MutableList<SqlParameter> = mutableListOf()
+    filters.filterNot { it.type == FilterType.BOOLEAN }.forEach { sqlParams.add(SqlParameter.builder().name(it.getKey()).value(it.value.lowercase()).build()) }
+    filters.filter { it.type == FilterType.BOOLEAN }.forEach { sqlParams.add(SqlParameter.builder().name(it.getKey()).value(it.value).build()) }
+    log.debug("SQL parameters: {}", sqlParams)
+    return sqlParams
+  }
+}
