+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/service/ConfiguredApiService.kt
@@ -8,6 +8,8 @@ import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.Configu
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.Count
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.ConfiguredApiRepository
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.ProductDefinitionRepository
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RedshiftDataApiRepository
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RepositoryHelper
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.Dataset
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.FilterDefinition
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.FilterType
@@ -23,6 +25,7 @@ import java.time.format.DateTimeParseException
 class ConfiguredApiService(
   val productDefinitionRepository: ProductDefinitionRepository,
   val configuredApiRepository: ConfiguredApiRepository,
+  val redshiftDataApiRepository: RedshiftDataApiRepository,
   @Value("\${URL_ENV_SUFFIX:#{null}}") val env: String? = null,
 ) {
 
@@ -84,7 +87,7 @@ class ConfiguredApiService(
     val productDefinition = productDefinitionRepository.getSingleReportProductDefinition(reportId, reportVariantId, dataProductDefinitionsPath)
     val dynamicFilter = buildAndValidateDynamicFilter(reportFieldId, prefix, productDefinition)
     val policyEngine = PolicyEngine(productDefinition.policy, userToken)
-    return configuredApiRepository
+    return redshiftDataApiRepository
       .executeQueryAsync(
         query = productDefinition.dataset.query,
         filters = validateAndMapFilters(productDefinition, filters) + dynamicFilter,
@@ -181,30 +184,30 @@ class ConfiguredApiService(
     return ConfiguredApiRepository.Filter(
       field = fieldId,
       value = prefix,
-      type = ConfiguredApiRepository.FilterType.DYNAMIC,
+      type = RepositoryHelper.FilterType.DYNAMIC,
     )
   }
 
-  private fun mapFilterType(filterDefinition: FilterDefinition, key: String, truncatedKey: String, schema: Schema): ConfiguredApiRepository.FilterType {
+  private fun mapFilterType(filterDefinition: FilterDefinition, key: String, truncatedKey: String, schema: Schema): RepositoryHelper.FilterType {
     if (filterDefinition.type == FilterType.DateRange) {
       if (key.endsWith(RANGE_FILTER_START_SUFFIX)) {
-        return ConfiguredApiRepository.FilterType.DATE_RANGE_START
+        return RepositoryHelper.FilterType.DATE_RANGE_START
       } else if (key.endsWith(RANGE_FILTER_END_SUFFIX)) {
-        return ConfiguredApiRepository.FilterType.DATE_RANGE_END
+        return RepositoryHelper.FilterType.DATE_RANGE_END
       }
     }
 
     if (key.endsWith(RANGE_FILTER_START_SUFFIX)) {
-      return ConfiguredApiRepository.FilterType.RANGE_START
+      return RepositoryHelper.FilterType.RANGE_START
     } else if (key.endsWith(RANGE_FILTER_END_SUFFIX)) {
-      return ConfiguredApiRepository.FilterType.RANGE_END
+      return RepositoryHelper.FilterType.RANGE_END
     }
     val schemaField = schema.field.first { it.name == truncatedKey }
     if (schemaField.type == ParameterType.Boolean) {
-      return ConfiguredApiRepository.FilterType.BOOLEAN
+      return RepositoryHelper.FilterType.BOOLEAN
     }
 
-    return ConfiguredApiRepository.FilterType.STANDARD
+    return RepositoryHelper.FilterType.STANDARD
   }
 
   fun findFilterDefinition(definition: SingleReportProductDefinition, filterName: String): FilterDefinition {
