+++ src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -12,4 +12,5 @@ uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.ReportDefinitionM
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.ReportDefinitionSummaryMapper
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.ReportDefinitionService
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.ConfiguredApiRepository
+uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.RedshiftDataApiRepository
 uk.gov.justice.digital.hmpps.digitalprisonreportinglib.config.DigitalPrisonReportingExceptionHandler
